refactor: move entity util
