+++ ee/clickhouse/queries/trends/trend_event_query.py
@@ -1,11 +1,11 @@
 from typing import Any, Dict, Tuple
 
-from ee.clickhouse.models.entity import get_entity_filtering_params
 from ee.clickhouse.models.property import get_property_string_expr
 from ee.clickhouse.queries.event_query import EnterpriseEventQuery
 from ee.clickhouse.queries.trends.util import get_active_user_params
 from posthog.constants import MONTHLY_ACTIVE, WEEKLY_ACTIVE, PropertyOperatorType
 from posthog.models import Entity
+from posthog.models.entity import get_entity_filtering_params
 from posthog.models.filters.filter import Filter
 from posthog.models.filters.mixins.utils import cached_property
 from posthog.models.utils import PersonPropertiesMode
